Converted EDGE_DISTANCE and EDGE_VELOCITY from const float to const unsigned short
(bzr r3472.3.4)

+++ plugins/wobbly/src/wobbly.cpp
@@ -29,10 +29,11 @@
 
 COMPIZ_PLUGIN_20090315 (wobbly, WobblyPluginVTable)
 
+
 const float MASS = 15.0f;
 
-const float EDGE_DISTANCE = 25.0f;
-const float EDGE_VELOCITY = 13.0f;
+const unsigned short EDGE_DISTANCE = 25;
+const unsigned short EDGE_VELOCITY = 13;
 
 void
 WobblyWindow::findNextWestEdge (Object *object)
